Moving into ModelBound structure
